[Refactor] Remove unused variable in _handle_dry_run
The 'all_run_cmds' variable in the _handle_dry_run method was initialized and appended to, but its value was never used.

This change removes the unused variable and its associated append operation to improve code clarity and remove dead code. This change does not alter the logic of the method, which is to print the generated run commands to the console.

Test: atest atest_unittests
OFF_PEAK_PRESUBMIT: true

Change-Id: I56bc55141dff446d3128f07f637849ae55478623