[Refactor] Move wait_for_debugger to _ARG_TO_CONST_MAP
Move wait_for_debugger from a special case in get_extra_args to the
_ARG_TO_CONST_MAP dictionary. This refactoring simplifies the get_extra_args
function by making it more declarative and data-driven. Centralizing the
argument mapping in one place improves readability and maintainability.

The original logic is preserved because the test runners only check for the
existence of the WAIT_FOR_DEBUGGER key in the extra_args dictionary, and
are not sensitive to its value (None vs. True). This change will continue
to add the key to extra_args when the --wait-for-debugger flag is used.

Test: atest atest_unittests
OFF_PEAK_PRESUBMIT: true

Change-Id: I956b16b54d230bae6a95284eb463329de92a2800