[Refactor] Move _StreamToLogger to module level
Moved the nested class  out of the  function to make it a module-level class.

This change improves code organization and readability by decoupling the class definition from the function that uses it. It follows the best practice of defining classes at the module level unless they are strictly helper classes that are only used within a single function and have no external relevance.

This refactoring has no impact on the functionality or logic of the program.

Test: atest atest_unittests
OFF_PEAK_PRESUBMIT: true

Change-Id: I4fc2dab3fa79bbe2e5a68049077ec0ffb0615912