[Refactor] Simplify custom_args handling in arg_parser
Set a default value for the 'custom_args' argument in the argparse setup.

By setting `default=[]` for the '--' argument, `parsed_args.custom_args` is guaranteed to be a list, even when the '--' separator is not present in the command-line arguments. This eliminates the need for the subsequent check `if not parsed_args.custom_args: parsed_args.custom_args = []` in the `parse_args` function.

This change simplifies the code by removing a redundant check and makes the argument definition more robust and explicit. It does not alter any program logic, as it simply moves the default value assignment into the argparse definition, which is the idiomatic way to handle default values for optional arguments.

Test: atest atest_unittests
OFF_PEAK_PRESUBMIT: true

Change-Id: I99b32fec3b8b9e116095b58117dcd8d709b55558