[Refactor] Move _BUILD_OUTPUT_MODE_DESCRIPTIONS to module level
Moving the _BUILD_OUTPUT_MODE_DESCRIPTIONS dictionary from within the _output_mode_msg function to a module-level constant. This improves clarity by signaling that the data is static and does not change with each function call. It also offers a minor efficiency improvement by preventing the dictionary from being recreated every time the function is invoked. This change does not alter the existing logic or behavior of the code.

Test: atest atest_unittests
OFF_PEAK_PRESUBMIT: true

Change-Id: I8f03828207732db6e84d9b93e2ccd97dacf74aeb