[Refactor] Use | None for optional type in atest_configs.py
Refactor GLOBAL_ARGS type hint to use `argparse.Namespace | None`
instead of `Optional[argparse.Namespace]`.

This change modernizes the code to use the preferred PEP 604 syntax for
optional types, which was introduced in Python 3.10. This improves
readability and conciseness.

The `typing.Optional` import is also removed as it is no longer
necessary.

This is a purely stylistic change and does not alter any logic or
program behavior.

Test: atest atest_unittests
OFF_PEAK_PRESUBMIT: true

Change-Id: I1a12a51d4cd196d05c3746ce7c2c6c9a713aef58