[Refactor] Standardize DetectType enum docstring
Update the docstring for the  enum to be more concise and consistent with common Python enum documentation practices. This change improves clarity and readability by making the docstring direct and focused on the purpose of the enum.

This refactoring is purely cosmetic and does not alter the existing logic, functionality, or behavior of the code. The enum members and their values remain unchanged, ensuring no impact on unit tests or any part of the Atest

Test: atest atest_unittests
OFF_PEAK_PRESUBMIT: true

Change-Id: I55103222b3b500d70a4f4cc7eca6763516d89936