[Refactor] Use list comprehension in append_test_info_to_invocation_pathnames
This change refactors a for loop into a list comprehension
in the append_test_info_to_invocation_pathnames function.

Using a list comprehension is more Pythonic and improves
the readability and conciseness of the code.

This is a purely stylistic change and does not alter the
existing logic or behavior of the code.

Test: atest atest_unittests
OFF_PEAK_PRESUBMIT: true

Change-Id: Ia6c785b1507111b7331d87527137d7a544c57827