[Refactor] Use self to call static method in instance method.
Calling a static method via  within an instance method is a common practice for consistency and readability, even though calling it via the class name is also valid. This change makes the code slightly more consistent in its access patterns within the instance.

This is a stylistic change and does not alter the logic or behavior of the program.

Test: atest atest_unittests
OFF_PEAK_PRESUBMIT: true

Change-Id: I8b227d903267cb9747f29255a7fc65ec77e3f883