[Refactor] [bug fix] Use 'or' for default values in __init__
This change refactors the __init__ method of AtestExecutionInfo
to use the 'or' operator for setting default values for _start_time
and _repo_out_dir.

This is a more idiomatic and concise way to assign a default value
in Python when the alternative is None.

This change does not alter the logic of the program.

Test: atest atest_unittests
OFF_PEAK_PRESUBMIT: true

Change-Id: Id900c63984adebb62d487ce7fe50f0562516e591