[Refactor] Use f-string for test_function_name in atest_integration_tests.py
This change updates the string formatting for 'test_function_name' in create_test_method to use an f-string instead of the old-style '%' formatting. This improves readability and aligns with modern Python best practices (PEP 498). The logic and functionality of the code remain unchanged, as it only alters the syntax of string interpolation. This refactoring does not impact any existing behavior or introduce any new side effects.

Test: atest atest_unittests
Flag: TEST_ONLY
OFF_PEAK_PRESUBMIT: true

Change-Id: Id4de4dc4c5f324835b247a7f1a7069b959274ac5