[Refactor] Use f-string for better readability in create_test_run_dir
This change replaces the old '%' style string formatting with an f-string in the 'create_test_run_dir' function. F-strings are generally preferred in modern Python for their conciseness and readability, especially when embedding expressions within string literals. This refactoring is purely stylistic and does not alter the existing logic or behavior of the code in any way.

Test: atest atest_unittests
Flag: TEST_ONLY
OFF_PEAK_PRESUBMIT: true

Change-Id: I970a75c0e7d24fb6fd27d9900768cdfd68578e32