[Refactor] Use mock.patch for better test isolation
Refactor the test to use @mock.patch decorator instead of direct assignment to mock a class.

Directly assigning a mock to a class in a module can lead to test pollution, where the mock persists after the test has finished, potentially affecting other tests. Using @mock.patch as a decorator or context manager ensures that the patch is automatically removed after the test execution.

This change does not alter the test's logic. It continues to verify that LocalDetectEvent is called with the correct parameters. It only improves the test's isolation and reliability.

Test: atest atest_unittests
Flag: TEST_ONLY
OFF_PEAK_PRESUBMIT: true

Change-Id: I445bc661835eeeb0ee03cdb4e85e5c1e781a8fa3