[Refactor] Use assertIsNotNone in atest_main_unittest
In test_make_test_run_dir, assertTrue(date_time) is used to check if a datetime object was successfully created.

This is being changed to assertIsNotNone(date_time) for better readability and to be more explicit about the intent of the check. The original check works because datetime objects are truthy, but assertIsNotNone more clearly conveys that the test is checking for the non-nullness of the object after the strptime call.

This change does not alter the logic as both assertions will pass if a datetime object is created and fail otherwise (though assertTrue would also fail for other "falsey" values, which is not possible for a datetime object).

Test: atest atest_unittests
Flag: TEST_ONLY
OFF_PEAK_PRESUBMIT: true

Change-Id: I0aed7dfa2e1c8e1b6b74b34df277e815044d5c31